+++ sklearn/cluster/k_means_.py
@@ -446,7 +446,8 @@ def _labels_inertia(X, x_squared_norms, centers,
         The cluster centers.
 
     distances: float64 array, shape (n_samples,)
-        Distances for each sample to its closest center.
+        Pre-allocated array to be filled in with each sample's distance
+        to the closest center.
 
     Returns
     -------
@@ -846,27 +847,26 @@ def _mini_batch_step(X, x_squared_norms, centers, counts,
     # Perform label assignment to nearest centers
     nearest_center, inertia = _labels_inertia(X, x_squared_norms, centers,
                                               distances=distances)
+
     if random_reassign and reassignment_ratio > 0:
         random_state = check_random_state(random_state)
         # Reassign clusters that have very low counts
         to_reassign = np.logical_or(
             (counts <= 1), counts <= reassignment_ratio * counts.max())
-        number_of_reassignments = to_reassign.sum()
-        if number_of_reassignments:
+        n_reassigns = min(to_reassign.sum(), X.shape[0])
+        if n_reassigns:
             # Pick new clusters amongst observations with probability
             # proportional to their closeness to their center.
             # Flip the ordering of the distances.
             distances -= distances.max()
             distances *= -1
-            rand_vals = random_state.rand(number_of_reassignments)
+            rand_vals = random_state.rand(n_reassigns)
             rand_vals *= distances.sum()
             new_centers = np.searchsorted(distances.cumsum(),
                                           rand_vals)
             if verbose:
-                n_reassigns = to_reassign.sum()
-                if n_reassigns:
-                    print("[MiniBatchKMeans] Reassigning %i cluster centers."
-                          % n_reassigns)
+                print("[MiniBatchKMeans] Reassigning %i cluster centers."
+                      % n_reassigns)
 
             if sp.issparse(X) and not sp.issparse(centers):
                 assign_rows_csr(X, new_centers, np.where(to_reassign)[0],
@@ -874,7 +874,7 @@ def _mini_batch_step(X, x_squared_norms, centers, counts,
             else:
                 centers[to_reassign] = X[new_centers]
 
-    # implementation for the sparse CSR reprensation completely written in
+    # implementation for the sparse CSR representation completely written in
     # cython
     if sp.issparse(X):
         return inertia, _k_means._mini_batch_update_csr(
@@ -941,7 +941,7 @@ def _mini_batch_convergence(model, iteration_idx, n_iter, tol,
     if verbose:
         progress_msg = (
             'Minibatch iteration %d/%d:'
-            'mean batch inertia: %f, ewa inertia: %f ' % (
+            ' mean batch inertia: %f, ewa inertia: %f ' % (
                 iteration_idx + 1, n_iter, batch_inertia,
                 ewa_inertia))
         print(progress_msg)