+++ sklearn/covariance/graph_lasso_.py
@@ -211,7 +211,7 @@ def graph_lasso(emp_cov, alpha, cov_init=None, mode='cd', tol=1e-4,
                                          'too ill-conditioned for this solver')
         else:
             warnings.warn('graph_lasso: did not converge after %i iteration:'
-                          'dual gap: %.3e' % (max_iter, d_gap),
+                          ' dual gap: %.3e' % (max_iter, d_gap),
                           ConvergenceWarning)
     except FloatingPointError as e:
         e.args = (e.args[0]
@@ -556,8 +556,8 @@ class GraphLassoCV(GraphLasso):
         # Finally, compute the score with alpha = 0
         alphas.append(0)
         grid_scores.append(cross_val_score(EmpiricalCovariance(), X,
-                                         cv=cv, n_jobs=self.n_jobs,
-                                         verbose=inner_verbose))
+                                           cv=cv, n_jobs=self.n_jobs,
+                                           verbose=inner_verbose))
         self.grid_scores = np.array(grid_scores)
         best_alpha = alphas[best_index]
         self.alpha_ = best_alpha