+++ sklearn/metrics/__init__.py
@@ -51,6 +51,8 @@ from .cluster import (adjusted_rand_score,
 
 from .pairwise import (euclidean_distances,
                        pairwise_distances,
+                       pairwise_distances_argmin_min,
+                       pairwise_distances_argmin,
                        pairwise_kernels)
 
 __all__ = ['accuracy_score',
@@ -64,6 +66,7 @@ __all__ = ['accuracy_score',
            'completeness_score',
            'confusion_matrix',
            'euclidean_distances',
+           'pairwise_distances_argmin_min',
            'explained_variance_score',
            'f1_score',
            'fbeta_score',