+++ sklearn/neighbors/tests/test_ball_tree.py
@@ -166,8 +166,8 @@ def test_ball_tree_kde(n_samples=100, n_features=3):
                 dens = bt.kernel_density(Y, h, atol=atol, rtol=rtol,
                                          kernel=kernel,
                                          breadth_first=breadth_first)
-                assert_allclose(dens, dens_true, atol=atol,
-                                          rtol=max(rtol, 1e-7))
+                assert_allclose(dens, dens_true,
+                                atol=atol, rtol=max(rtol, 1e-7))
 
             for rtol in [0, 1E-5]:
                 for atol in [1E-6, 1E-2]: